Update pco.py
Make it possible to wait longer for the very first camera trigger in a series of exposures written to a single buffer.

Very large voltage arrays to be loaded to the NI daq card were taking >~1 second to load, and that exceeded the camera timeout defined by previous versions of pco.py. Those timeouts were initially defined to handle missed triggers in a series of exposures being written to a single buffer.