Update pco.py
The roi legalizer has been split out from the Edge class as separate function so it can be called without instantiating the camera. The reason for this is that image data pipeline will assign the image buffer before instantiating the camera (which sets the roi) but the roi may not be legal. This helps ensure the buffer created is the correct size and not undersized which will crash.