Changed digital out api. Support for external clock. Supports NI 9401.
The api change is that each instance is now only analog or only digital. This is inferred by the daq_type argument. 

I have not tested this on the analog or digital lines of the NI6733.  Digital lines of NI6733 are accessed by daq_type='6733_digital'. With the NI6733 it may be necessary start/stop the digital voltage plays from the analog clock.