BREAKING CHANGE | Added support for  pixelfly
Original code was written with only pco.Edge cameras in mind. While adding support of pixelfly, it became obvious that class names and keyword arguments should change. The logic remains the same. 

Any code that depends on this should be trivial to modify to work with this version, however it will not run out-of-the-box.