Restores functionality of 6733 digital channels
The ni6733 DO channels do not have their own clock and the AO sample clock must be specified ('/Dev1/ao/SampleClock") when issuing the `set_rate` dll call . Previous version of this code assumed that this DLL call during initialization could be made without specifying a sample clock (defaulting to None), but this fails on the 6733. 

See comment in the `__init__` function, but this fix slightly changes the behavior of how the 6733_digital is initialized and it is best practice initialize the 6733_digital card immediately prior to initializing the  6733 (analog).