Update np_tif.py
Added an experimental function to run imagej macros on numpy arrays. It's clunky (saves to disk, loads from disk), but it's not so complicated, and it's eases a frequent pain point when an ImageJ plugin does something you don't want to re-implement in python. There's a half-assed tkinter GUI dialog to walk you through finding ImageJ the first time.