Update proxy_objects.py
np.prod() defaults to using the dtype of the 1st argument which is surprisingly small bitdepth for Python ints! When multiplying large numbers this can easily result in nonsense output (see docs at: https://numpy.org/doc/stable/reference/generated/numpy.prod.html)

For example: Arithmetic is modular when using integer types, and no error is raised on overflow. That means that, on a 32-bit platform:
>>> x = np.array([536870910, 536870910, 536870910, 536870910])
>>> np.prod(x)
16 # may vary