created threads that crash on join
Sometimes we had tasks running in threads that would raise an Exception at some frequency. Previously when this occurred in proxied objects with custody locks this would result in a silent hang.

To address this, we implemented threads (_RaiseOnJoinThread) that will raise any exception that occured when their join method is called.