locate SC2_Cam.dll in the directory with pco.py
The previous behavior forced SC2_Cam.dll to be located in the current working directory. This was annoying and forced an organization that might not work with all projects. 

A while ago (prior to c0c33260ee6ba08df50745ce0d9bf80fad256260), SC2_Cam.dll just had to be located on the system PATH, however new versions of python/ctypes changed the way they located DLLs and forced a more specific path to be specified.  

This change allows SC2_Cam.dll to be located in the same folder as pco.py, independent of the current working directory. As long as pco.py is on the PATH/PYTHONPATH, it should be able build the correct path to load the DLL. I think this is closer to the original behavior.