Update picoquant_tttr.py
I added two changes to the header parsing function based on our conversation on 11/30. I moved the comment from line 35 to line 37. I also switched the sign on line 66 so that booleans don't switch signs when they import. I haven't tested that the behavior with booleans is correct with LSM_1.ptu yet.